+++ crates/loro-internal/src/state/container_store.rs
@@ -161,6 +161,11 @@ impl ContainerStore {
             .with_container_for_read(idx, |c| c.text_utf16_len(idx, ctx!(self)))?
     }
 
+    pub fn text_utf8_len(&mut self, idx: ContainerIdx) -> Option<usize> {
+        self.store
+            .with_container_for_read(idx, |c| c.text_utf8_len(idx, ctx!(self)))?
+    }
+
     pub fn has_decoded_state(&mut self, idx: ContainerIdx) -> bool {
         self.store.has_decoded_state(idx)
     }