+++ crates/loro-internal/tests/mergeable_container.rs
@@ -9,7 +9,11 @@ mod delete;
 mod discriminator;
 #[path = "mergeable_container/events_and_paths.rs"]
 mod events_and_paths;
+#[path = "mergeable_container/pending.rs"]
+mod pending;
 #[path = "mergeable_container/snapshot.rs"]
 mod snapshot;
 #[path = "mergeable_container/type_conflict.rs"]
 mod type_conflict;
+#[path = "mergeable_container/undo_redo.rs"]
+mod undo_redo;